% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abstractsHostParasites-package.R
\docType{package}
\name{abstractsHostParasites-package}
\alias{abstractsHostParasites}
\alias{abstractsHostParasites-package}
\title{abstractsHostParasites: Literature Abstract Classification System}
\description{
Create abstract classification models with PU learning approach using a set of abstract of an specific class and random abstracts labeled as unknown class. Also helps to get random abstracts from connecting to the 'CrossRef' 'API'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/alrobles/abstractsHostParasites}
  \item Report bugs at \url{https://github.com/alrobles/abstractsHostParasites/issues}
}

}
\keyword{internal}
