% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automodel_pu_abstracts.R
\name{automodel_pu_abstracts}
\alias{automodel_pu_abstracts}
\title{Automodel is a function to create an abstract classfication
using a data frame with abstracts and two given target classes.
One of this is labeled as unknown. The model is carried out with
an implementation of the PU learning algorithm. (You can find information
of the algorithm in the function PLUS)}
\usage{
automodel_pu_abstracts(abstracts_data_frame)
}
\arguments{
\item{abstracts_data_frame}{A data frame with abstracts and class of abstracts
to be classified using PU learning via PLUS function}
}
\value{
a list of 2 objects. A model and a vocabulary associated to that model.
}
\description{
Automodel is a function to create an abstract classfication
using a data frame with abstracts and two given target classes.
One of this is labeled as unknown. The model is carried out with
an implementation of the PU learning algorithm. (You can find information
of the algorithm in the function PLUS)
}
\examples{
\dontrun{
data(df_abstracts_sample)
output <- automodel_pu_abstracts(df_abstracts_sample)
#' }
}
