% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_abstracts_random.R
\docType{data}
\name{df_abstracts_random}
\alias{df_abstracts_random}
\title{Data frame with 4236 random abstracts retrived from crossref.}
\format{
An object of class \code{"data.frame"};
}
\usage{
data(df_abstracts_random)
}
\description{
Data frame with 4236 random abstracts retrived from crossref.
}
\examples{
data(label_obs)
\donttest{df_abstracts_random}
}
\keyword{datasets}
