% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_abstracts_random_sample.R
\docType{data}
\name{df_abstracts_random_sample}
\alias{df_abstracts_random_sample}
\title{Data frame with 800 random abstracts retrived from crossref.}
\format{
An object of class \code{"data.frame"};
}
\usage{
data(df_abstracts_random_sample)
}
\description{
Data frame with 800 random abstracts retrived from crossref.
}
\examples{
data(df_abstracts_random_sample)
\donttest{df_abstracts_random_sample}
}
\keyword{datasets}
