% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_abstracts_sample.R
\docType{data}
\name{df_abstracts_sample}
\alias{df_abstracts_sample}
\title{Data frame with 2394 random abstracts. Each abstract belongs to
one of both classes, parasite and unknown. Abstracts from parasite class
are from ZOVER and GMPD database. Abstracs from uknown class are retrived
from crossref.}
\format{
An object of class \code{"data.frame"};
}
\usage{
data(df_abstracts_sample)
}
\description{
Data frame with 2394 random abstracts. Each abstract belongs to
one of both classes, parasite and unknown. Abstracts from parasite class
are from ZOVER and GMPD database. Abstracs from uknown class are retrived
from crossref.
}
\examples{
data(label_obs)
\donttest{df_abstracts_sample}
}
\keyword{datasets}
