% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtm_df_tfidf_train.R
\docType{data}
\name{dtm_df_tfidf_train}
\alias{dtm_df_tfidf_train}
\title{Document term matrix with dt - tfidf matrix to train an
abstract classification using positive and unlabel learning}
\format{
An object of class \code{"Matrix"};
}
\usage{
data(dtm_df_tfidf_train)
}
\description{
Document term matrix with dt - tfidf matrix to train an
abstract classification using positive and unlabel learning
}
\examples{
data(dtm_df_tfidf_train)
\donttest{dim(dtm_df_tfidf_train)}
}
\keyword{datasets}
