% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_class_score.R
\name{get_class_score}
\alias{get_class_score}
\title{get_class_score}
\usage{
get_class_score(model, vocabulary, abstract)
}
\arguments{
\item{model}{A PU learning model. Can provide from
automodel_pu_abstracts}

\item{vocabulary}{A vocabulary asociated to the PU learning model.}

\item{abstract}{A long string with the abstract to be evaluated}
}
\value{
A numeric value from 0 to 1. Is interpreted as the probability
of belonging to the class of the data set modeled.
}
\description{
Function to get a classification score given an abstract
classification model and vocabulary associated with the model.
Evaluates a string (abstracts) and returs the probability to belong a class
}
\examples{
library(dplyr)
output <- automodel_pu_abstracts(df_abstracts_sample, term_count = 20)
df_abstracts_score <- df_abstracts_sample \%>\%
  dplyr::slice(1:100) \%>\%
  dplyr::select(abstract, class) \%>\%
  dplyr::mutate(score = purrr::map_dbl(abstract, function(x){
      get_class_score(
      model = output$model$fit.pi,
      vocabulary =  output$vocabulary,
      abstract = x)
      }) )
}
