% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_obs.R
\docType{data}
\name{label_obs}
\alias{label_obs}
\title{Labeled observations associated to document term matrix
with dt - tfidf matrix to train an abstract classification model
using positive and unlabel learning}
\format{
An object of class \code{"Matrix"};
}
\usage{
data(label_obs)
}
\description{
Labeled observations associated to document term matrix
with dt - tfidf matrix to train an abstract classification model
using positive and unlabel learning
}
\examples{
data(label_obs)
\donttest{table(label_obs)}
}
\keyword{datasets}
