% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_fun.R
\name{prep_fun}
\alias{prep_fun}
\title{prep_fun}
\usage{
prep_fun(x)
}
\arguments{
\item{x}{String with an abstract (or title) to preprocess}
}
\value{
A processed string
}
\description{
Preprocesing function. Change all characters to lower case
replace the alfanumerics and remove the spaces and the line breaks
}
\examples{
\dontrun{
prep_fun("This is an Example  of the string")
}
}
