% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_abstract.R
\name{search_abstract}
\alias{search_abstract}
\title{Search Abstrac}
\usage{
search_abstract(doi = NULL, title = NULL, email = NULL, APIKEY = NULL)
}
\arguments{
\item{doi}{A valid doi}

\item{title}{A title for search if does not have a doi}

\item{email}{An email for the pool in the crossref API}

\item{APIKEY}{An entrez API KEY}
}
\value{
A data frame with abstracts
}
\description{
This function search an abstract in crossref
(if there is any DOI) or search in PubMed with entrez API
}
\examples{
#' \dontrun{
doi <- "10.3389/fvets.2021.604560"
title <- "American Mammals Susceptibility to Dengue"
search_abstract(
doi,
title
)
}
}
