% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_abstract_crossref.R
\name{search_abstract_crossref}
\alias{search_abstract_crossref}
\title{search_abstract_crossref}
\usage{
search_abstract_crossref(doi = NULL, mailto = NULL)
}
\arguments{
\item{doi}{A vaild doi to search in cross reference api}

\item{mailto}{A valid mail to the polite pool}
}
\value{
a Data frame with doi, title, abstract, year, month and day
}
\description{
search_abstract_crossref
}
\examples{
\dontrun{
doi <- "10.3389/fvets.2021.604560"
mail <- "your@example.com"
search_abstract_crossref(doi, mail)
}
}
