% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_abstract_pubmed.R
\name{search_abstract_pubmed}
\alias{search_abstract_pubmed}
\title{search_abstract_with_rentrez}
\usage{
search_abstract_pubmed(DOI = NULL, TITLE, APIKEY = NULL)
}
\arguments{
\item{DOI}{A valid DOI}

\item{TITLE}{A title to validate the search}

\item{APIKEY}{A entrez API key}
}
\value{
a data frame with doi title abstract
}
\description{
search_abstract_with_rentrez
}
\examples{
\dontrun{
doi <- "10.3389/fvets.2021.604560"
title <- "American Mammals Susceptibility to Dengue"
search_abstract_pubmed(
doi,
title
)
}
}
