% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abstractsHostParasites-package.R
\docType{package}
\name{abstractsHostParasites-package}
\alias{abstractsHostParasites}
\alias{abstractsHostParasites-package}
\title{abstractsHostParasites: What the Package Does (Title Case)}
\description{
This packages helps to create datasets of abstracts and create a model with a PU learning aproach that predicts if an abstract belongs to an specific class.
}
\keyword{internal}
