% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automodel_pu_abstracts.R
\name{automodel_pu_abstracts}
\alias{automodel_pu_abstracts}
\title{Automodel is a function to create an abstract classfication
using a data frame with abstracts and two given target classes.
One of this is labeled as unknown. The model is carried out with
an implementation of the PU learning algorithm. (You can find information
of the algorithm in the function PLUS)}
\usage{
automodel_pu_abstracts(
  abstracts_data_frame,
  term_count = 2,
  objetive_class = "parasite",
  split_prop = 0.65,
  doc_prop_min = 0.1,
  doc_prop_max = 0.5
)
}
\arguments{
\item{abstracts_data_frame}{A data frame with abstracts and class of abstracts
to be classified using PU learning via PLUS function}

\item{term_count}{Minimum frequency a word should have in each document}

\item{objetive_class}{string with the label of the target class of abstracts
to classify}

\item{split_prop}{proportion of the training and testing data}

\item{doc_prop_min}{Minimal proportion of each term in document term matrix
with respect to the document to be considered for modeling.}

\item{doc_prop_max}{Maximum proportion of each term in document term matrix
with respect to the document to be considered for modeling.}
}
\value{
a list of 2 objects. A model and a vocabulary associated
to that model.
}
\description{
Automodel is a function to create an abstract classfication
using a data frame with abstracts and two given target classes.
One of this is labeled as unknown. The model is carried out with
an implementation of the PU learning algorithm. (You can find information
of the algorithm in the function PLUS)
}
\examples{
\dontrun{
data(df_abstracts_sample)
output <- automodel_pu_abstracts(df_abstracts_sample)
}
}
